Cleanups and remember to do location permissions prompt.
